Refactor: 유틸 함수 파일명 변경
